# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sample API Spec for MyGram",
        "title": "MyGram API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "zayyan",
            "email": "zayyanramadhan@gmail.com"
        },
        "version": "v2.0"
    },
    "host": "localhost:7000",
    "basePath": "/",
    "paths": {
...
        "/socialmedias": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socialmedia"
                ],
                "summary": "Get SocialMedia",
                "responses": {}
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socialmedia"
                ],
                "summary": "Create SocialMedia",
                "parameters": [
                    {
                        "description": "SocialMedia info",
                        "name": "SocialMedia",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SocialMediaRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/socialmedias/{socialMediaId}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socialmedia"
                ],
                "summary": "Update SocialMedia",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "socialMediaId",
                        "name": "socialMediaId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "SocialMedia info",
                        "name": "SocialMedia",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SocialMediaRequest"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socialmedia"
                ],
                "summary": "Delete SocialMedia",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "socialMediaId",
                        "name": "socialMediaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/users/{userId}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "user info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserUpdate"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "models.SocialMediaRequest": {
            "type": "object",
            "required": [
                "name",
                "social_media_url"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "zayyan"
                },
                "social_media_url": {
                    "type": "string",
                    "example": "zayyan"
                }
            }
        },
...
        "models.UserUpdate": {
            "type": "object",
            "required": [
                "email",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "zayyan@mail.com"
                },
                "username": {
                    "type": "string",
                    "example": "zayyan"
                }
            }
        }
    }
}